Added ConnectionType and SystemType api models
