+++ clients/admin-ui/src/features/datastore-connections/datastore-connection.slice.ts
@@ -2,10 +2,12 @@ import { createSlice, PayloadAction } from "@reduxjs/toolkit";
 import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
 import { addCommonHeaders } from "common/CommonHeaders";
 
+import { SystemType } from "~/types/api";
+
 import type { RootState } from "../../app/store";
 import { BASE_URL, CONNECTION_ROUTE } from "../../constants";
 import { selectToken } from "../auth";
-import { DisabledStatus, SystemType, TestingStatus } from "./constants";
+import { DisabledStatus, TestingStatus } from "./constants";
 import {
   CreateAccessManualWebhookRequest,
   CreateAccessManualWebhookResponse,
