+++ clients/admin-ui/src/features/datastore-connections/types.ts
@@ -1,10 +1,10 @@
+import { ConnectionType, SystemType } from "~/types/api";
+
 import {
   AccessLevel,
   ConnectionTestStatus,
-  ConnectionType,
   DisabledStatus,
   SaasType,
-  SystemType,
   TestingStatus,
 } from "./constants";
