+++ clients/admin-ui/src/types/api/models/SystemType.ts
@@ -1,0 +1,12 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+/**
+ * An enumeration.
+ */
+export enum SystemType {
+  SAAS = "saas",
+  DATABASE = "database",
+  MANUAL = "manual",
+}
